UH03 | Refactor headers package
