--- src/test/java/starter/pages/articleslist/ArticlesListContent.java
@@ -1,8 +1,0 @@
-package starter.pages.articleslist;
-
-import net.serenitybdd.screenplay.targets.Target;
-
-public class ArticlesListContent {
-    public static final Target HEADER_TITLE = Target.the("title").locatedBy(".title");
-
-}
