+++ src/test/java/starter/pages/cartdetail/CartDetailContent.java
@@ -1,4 +1,7 @@
 package starter.pages.cartdetail;
 
+import net.serenitybdd.screenplay.targets.Target;
+
 public class CartDetailContent {
+    public static final Target HEADER_TITLE = Target.the("title").locatedBy(".title");
 }
