+++ src/test/java/starter/pages/headers/HeadersContent.java
@@ -1,7 +1,9 @@
-package starter.pages.primaryheader;
+package starter.pages.headers;
 
 import net.serenitybdd.screenplay.targets.Target;
 
-public class PrimaryHeaderContent {
+public class HeadersContent {
+
+    public static final Target HEADER_TITLE = Target.the("title").locatedBy(".title");
     public static final Target CART_ICON = Target.the("cart icon").locatedBy(".shopping_cart_link");
 }
